Run script with sudo privileges
The current approach is not very nice because, during the
installation of some packages, we lose sudo privileges and
require to type the password again.

And if I am not paying attention everytime the script crashes.

I changed it so it can be run with `sudo` directly.

Of course, commands that don't need it are run with my regular user
instead.